Move the _check_sigs_and_hash_and_fetch from FederationBase to FederationClient.
